# client/src/main/resources/openapi/resos-openapi.yml
openapi: 3.0.0
info:
  title: ResOS API
  version: '1.2'
  description: |
    ResOS API for managing restaurant bookings, tables, opening hours, orders, feedback and customers.
    
    For support or feature requests, contact: hi@resos.com
  contact:
    email: hi@resos.com
servers:
  - url: https://api.resos.com/v1
    description: Production API Server
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using API key as username (no password required).
        The API key must be base64 encoded with a trailing colon.
  schemas:
...
    Order:
      type: object
      properties:
        _id:
          type: string
        bookingId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
              notes:
                type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  parameters:
...
    limit:
      name: limit
      in: query
      description: Number of records to return (max 100)
      required: false
      schema:
        type: integer
        default: 100
        maximum: 100
    skip:
      name: skip
      in: query
      description: Number of records to skip
      required: false
      schema:
        type: integer
        default: 0
    sort:
      name: sort
      in: query
      description: Sort field and direction (1 ascending, -1 descending)
      required: false
      schema:
        type: string
        example: "createdAt:1"
    customQuery:
      name: customQuery
      in: query
      description: Search expression for filtering results
      required: false
      schema:
        type: string
        example: "people>1,duration>60"
paths:
...
  /orders:
    get:
      summary: List orders
      description: Retrieve a list of orders
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/customQuery'
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get order
      description: Retrieve a specific order by ID
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
...